// FRONTEND-VENTA/src/app/core/models/enums.ts
export enum Role {
  ADMIN = 'ADMIN',
  SELLER = 'SELLER',
  VIEWER = 'VIEWER'
}

export enum SaleStatus {
  PENDING = 'PENDING',
  COMPLETED = 'COMPLETED',
  CANCELLED = 'CANCELLED'
}

export enum ProformaStatus {
  DRAFT = 'DRAFT',
  SENT = 'SENT',
  APPROVED = 'APPROVED',
  REJECTED = 'REJECTED',
  CONVERTED = 'CONVERTED'
}